rl-agent-comparison-louw/pyproject.toml
[project]
name = "rl-agent-comparison-louw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium>=1.2.3",
    "numpy>=2.4.0",
    "pyyaml>=6.0",
    "stable-baselines3>=2.7.1",
    "tensorboard>=2.18",
    "torch>=2.9.1",
]
